move httpserver to own pkg
