+++ pkg/httpserver/signal/signal.go
@@ -22,8 +22,8 @@ import (
 	"sync"
 	"syscall"
 
-	"github.com/trickstercache/trickster/v2/cmd/trickster/config"
 	"github.com/trickstercache/trickster/v2/pkg/cache"
+	"github.com/trickstercache/trickster/v2/pkg/config"
 	tl "github.com/trickstercache/trickster/v2/pkg/observability/logging"
 )
